+++ src/pages/_app.js
@@ -22,19 +22,6 @@ const anuphan = Anuphan({
 export default function App({ Component, pageProps }) {
   return (
     <>
-      <Script
-        strategy="lazyOnload"
-        src={`https://www.googletagmanager.com/gtag/js?id=G-Z5T1MSDEG5`}
-      ></Script>
-      <Script strategy="lazyOnload" id="public-ga">
-        {`
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-
-            gtag('config', 'G-Z5T1MSDEG5');
-          `}
-      </Script>
       <Head>
         <title>TalentLink | BKK</title>
         <meta
