+++ src/pages/_document.js
@@ -1,10 +1,44 @@
 import { Head, Html, Main, NextScript } from 'next/document';
+import Script from 'next/script';
 
 export default function Document() {
   return (
     <Html lang="en">
-      <Head />
+      <Head>
+        <Script
+          strategy="lazyOnload"
+          src={`https://www.googletagmanager.com/gtag/js?id=G-Z5T1MSDEG5`}
+        ></Script>
+        <Script id="public-ga" strategy="lazyOnload">
+          {`
+            window.dataLayer = window.dataLayer || [];
+            function gtag(){dataLayer.push(arguments);}
+            gtag('js', new Date());
+
+            gtag('config', 'G-Z5T1MSDEG5');
+          `}
+        </Script>
+        <Script
+          id="g-tag-mgr"
+          strategy="afterInteractive"
+          dangerouslySetInnerHTML={{
+            __html: `
+            (function(w,d,s,l,i){w[l]=w[l]||[];w[l].push({'gtm.start':
+            new Date().getTime(),event:'gtm.js'});var f=d.getElementsByTagName(s)[0],
+            j=d.createElement(s),dl=l!='dataLayer'?'&l='+l:'';j.async=true;j.src=
+            'https://www.googletagmanager.com/gtm.js?id='+i+dl;f.parentNode.insertBefore(j,f);
+            })(window,document,'script','dataLayer','GTM-NV3HD52');
+        `,
+          }}
+        ></Script>
+      </Head>
       <body>
+        <noscript
+          dangerouslySetInnerHTML={{
+            __html: `iframe src="https://www.googletagmanager.com/ns.html?id=GTM-NV3HD52"
+          height="0" width="0" style="display:none;visibility:hidden"></iframe>`,
+          }}
+        ></noscript>
         <Main />
         <NextScript />
       </body>
